+++ fibber_test.go
@@ -6,24 +6,24 @@ import (
 	"testing"
 )
 
-var fibSeq = []int64{0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597, 2584, 4181, 6765, 10946, 17711, 28657, 46368, 75025, 121393, 196418, 317811}
+var fibSeq = []int{0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597, 2584, 4181, 6765, 10946, 17711, 28657, 46368, 75025, 121393, 196418, 317811}
 var fibBigInts = make([]*big.Int, len(fibSeq))
 
 func init() {
 	for idx, i := range fibSeq {
-		fibBigInts[idx] = big.NewInt(i)
+		fibBigInts[idx] = big.NewInt(int64(i))
 	}
 }
 
 func TestFibOf(t *testing.T) {
-	checkFibAt(t, func(idx int) *big.Int { return Of(uint64(idx)) })
+	checkFibAt(t, func(idx int) *big.Int { return Of(uint(idx)) })
 }
 
 func TestFibForEach(t *testing.T) {
 	// Create an empty array that we can fill-in in our callback
 	forEachedGenerated := make([]*big.Int, len(fibSeq))
 	currentIdx := 0
-	ForEach(uint64(len(fibSeq)-1), func(i *big.Int) {
+	ForEach(uint(len(fibSeq)-1), func(i *big.Int) {
 		forEachedGenerated[currentIdx] = i
 		currentIdx++
 	})
@@ -32,7 +32,7 @@ func TestFibForEach(t *testing.T) {
 
 func TestCachedFib(t *testing.T) {
 	fibGen := NewMemoed()
-	checkFibAt(t, func(idx int) *big.Int { return fibGen.Of(uint64(idx)) })
+	checkFibAt(t, func(idx int) *big.Int { return fibGen.Of(uint(idx)) })
 	if fib50 := fibGen.Of(50); fib50.Cmp(big.NewInt(12586269025)) != 0 {
 		err := fmt.Sprintf("Fib of 50 was expected to be %v but got %v", 12586269025, fibGen.Of(50))
 		t.Error(err)
@@ -66,7 +66,7 @@ func ExampleOf() {
 func ExampleForEach() {
 	forEachedGenerated := make([]*big.Int, len(fibSeq))
 	currentIdx := 0
-	ForEach(uint64(len(fibSeq)-1), func(i *big.Int) {
+	ForEach(uint(len(fibSeq)-1), func(i *big.Int) {
 		forEachedGenerated[currentIdx] = i
 		currentIdx++
 	})
